Add workaround for migration of vms in the lab
Due to the fact that lab environment is deployed from the images of the
preprovisioned hosts, it has stored wrong ssh host keys on compute nodes
and when nova-compute wants to SSH from one compute to another, it fails
due to the "WARNING: REMOTE HOST IDENTIFICATION HAS CHANGED!" error.

Actually in this case it is expected that host identification will be
changed so to avoid issues like that, lets simply add small workaround
by adding "StrictHostKeyChecking no" to the ssh config file in the
nova_compute container.

Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>